image_manager: shadows on multi image display
